fix: 本番Sentryエラー3件を修正
- UniqueViolation: add_scan_keyword を INSERT ON CONFLICT DO NOTHING に変更
- ROUND(double precision, int): 2引数ROUND に ::numeric キャストを追加 (6箇所)
- SSL connection closed: psycopg2 に TCP keepalive オプションを追加、commit/rollback に再接続リトライを追加

+++ agents/research_agent.py
@@ -175,7 +175,7 @@ class ResearchAgent:
             SELECT p.product_name, p.platform as buy_platform,
                    l.selling_platform, p.purchase_price,
                    s.sale_price, s.net_profit,
-                   ROUND(s.net_profit / s.sale_price * 100, 1) as profit_rate
+                   ROUND((s.net_profit / s.sale_price * 100)::numeric, 1) as profit_rate
             FROM sales s
             JOIN listings l ON s.listing_id = l.id
             JOIN purchases p ON l.purchase_id = p.id
